Fix accept issue
When using ab to test the hello-world server. There is a great change that
one of the concurrent connections are not accepted, so the client would
time out. The new connections will cause a SIGPIPE. This is because the old
connection(s) are not correctly accepted by the server. The right way to do
accept in event-driven server is continuously accept new connections until
EWOULDBLOCK/EAGAIN.
This commit fixes this issue.